config: Add config symlinks to Microsoft/
We have previously been adding updated WSL x86 and arm64 configs to
Microsoft/ for WSL builds. Updated configs are now kept under:

- arch/x86/configs/config-wsl
- arch/arm64/configs/config-wsl-arm64

We want to keep the Microsoft/ directory and config files in their
current location for easy use to users. However, to avoid maintaining
updated configs in two locations and make sure all locations are always
up to date, change the configs under Microsoft/ to be symlinks pointing
to the new official config location.

Note: when building a kernel with the Microsoft/config-wsl*:

$ make KCONFIG_CONFIG=Microsoft/config-wsl

This will create a new Microsoft/config-wsl which is not a symlink. The
symlink will be moved to Microsoft/config-wsl.old which will still be
pointing to arch/x86/configs/config-wsl

Signed-off-by: Kelsey Steele <kelseysteele@microsoft.com>